Lots of little FF updates and added Google Analytics

+++ SamSmithNZ2017/SamSmithNZ2017.DB/dbo/Tables/ff_album.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TABLE [dbo].[ff_album] (
-    [album_key]          SMALLINT      NOT NULL,
+    [album_key]          INT      NOT NULL,
     [album_name]         VARCHAR (100) NULL,
     [album_release_date] SMALLDATETIME NULL,
     [album_label]        VARCHAR (50)  NULL, --Deprocating, not useful
