+++ SamSmithNZ2017/SamSmithNZ2017.DB/dbo/Tables/ff_show_song.sql
@@ -1,8 +1,8 @@
 ﻿CREATE TABLE [dbo].[ff_show_song] (
-    [showtrack_key]   SMALLINT NOT NULL,
-    [show_key]        SMALLINT NULL,
-    [song_key]        SMALLINT NULL,
-    [show_song_order] SMALLINT NULL,
+    [showtrack_key]   INT NOT NULL,
+    [show_key]        INT NULL,
+    [song_key]        INT NULL,
+    [show_song_order] INT NULL,
     CONSTRAINT [PK_ff_show_track] PRIMARY KEY CLUSTERED ([showtrack_key] ASC)
 );
